+++ src/providers/index.ts
@@ -5,7 +5,7 @@ export * from './drafts'
 export * from './user'
 export * from './joystream'
 export * from './connectionStatus'
-export * from './editVideoSheet'
+export * from './editVideoSheet/editVideoSheet'
 export * from './uploadsManager'
 export * from './transactionManager'
 export * from './dialogs'
